+++ CHANGES.md
@@ -12,7 +12,7 @@ Beta Releases
 * Added `fromCssColorString` to `Color` to create a `Color` instance from any CSS value.
 * Added `fromHsl` to `Color` to create a `Color` instance from H, S, L values.
 * Added `Scene.backgroundColor`.
-* Added `MeshFilters.encodePosition`.
+* Added `MeshFilters.encodePosition` and `EncodedCartesian3.encode`.
 * Fixed jitter artifacts with polygons.
 * Fixed camera tilt close to the `minimumZoomDistance`.
 * Fixed a bug that could lead to blue tiles when zoomed in close to the North and South poles.